pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "staged-primitives"
authors = [
    {name = "Pedro Rivero", email = "pedro.rivero@ibm.com"},
]
maintainers = [
    {name = "Pedro Rivero", email = "pedro.rivero@ibm.com"},
]
keywords = [
    "quantum computing",
    "primitives",
    "qiskit",
    "IBM",
]
readme = "README.md"
license = {file = "LICENSE.txt"}
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Physics",
]
dynamic = ["version", "description"]
requires-python = ">=3.8"
dependencies = [
    "qiskit-terra >= 0.22.2",
]

[project.optional-dependencies]
dev = [
    "tox >= 3.25.1",
    "pre-commit >= 2.19.0",
    "commitizen >= 2.28.0",
]
test = [
    "pytest >= 7.1.2",
    "pytest-cov >= 3.0.0",
    "pytest-randomly >= 3.12.0",
]
lint = [
    "autoflake >= 1.7.6",
    "black[jupyter] ~= 22.6.0",
    "flake8 >= 4.0.1",
    "isort >= 5.10.1",
    "mypy >= 0.961",
    "mypy-extensions >= 0.4.3",
    "pylint >= 2.14.4",
]
docs = [
    "jupyter-sphinx >= 0.3.2",
    "nbsphinx >= 0.8.8",
    "reno >= 3.5.0",
    "sphinx-autodoc-typehints >= 1.17.0",
]
notebook = [
    "jupyter >= 1.0.0",
    "notebook >= 6.4.12",
    "qiskit-terra[visualization] >= 0.22.2",
    "qiskit-aer >= 0.11.0",
    "nbqa >= 1.3.1",
    "treon >= 0.1.3",
]

[project.urls]
"Home" = "https://qiskit-community.github.io/prototypes/"
"Documentation" = "https://github.com/Qiskit-Extensions/staged-primitives/tree/main/docs/"
"Repository" = "https://github.com/Qiskit-Extensions/staged-primitives/"
"Download" = "https://pypi.org/project/staged-primitives/"

[tool.flit.module]
name = "staged_primitives"


############################################################
## TOOLS
############################################################
[tool.black]
line-length = 100
exclude = '\.ipynb$'
extend-exclude = '.*/qiskit/.*'
